kustos: Execute enforce in blocking thread
Enforce can take a longe time to complete and locks up the async runtime.
Moving it into a blocking thread yields a _significant_ improvements to
the amount of concurrent requests served. It blocks from a few milliseconds
up to many seconds under high load with a large set of policies. (Thats a
separate issue to be investigated).