kustos: Avoid using regex in ACL model when enforcing
This _greatly_ increases performance when enforcing our ACL rules
since it avoids creating a regex _every single time_ casbin tries
to match the object and action.

Naturally this also limits how expressive we can be when building
policies, but I believe if we want to express more complex policies
we should implement custom and efficient matching functions instead
of relying on regexes.