logging: `RUST_LOG` environment variable entries override settings from configuration file
Changes the configuration of logging:
- Load logging information from `RUST_LOG` environment variable
- If no `RUST_LOG` environment variable is set, use a default defined by the `controller`
- Load the log filters from the configuration file, but only set them if they were not
  yet configured

In order to implement that behavior, we removed the `Default` implementation for
the `Logging` struct in `settings`, and replaced some of the default values by
`Option`s. Otherwise we wouldn't be able to determine whether the values were
set by the user or came from the default values.

Co-authored-by: Max Allies <m.allies@opentalk.eu>

Closes: #69