ci: extract code coverage
The code coverage extraction uses instrumentation through the
`-C instrument-coverage` parameter for `rustc`.

Recorded data is extracted by `grcov` afterwards.

Because the CI runner would otherwise kill the linker process due to
exhaustive memory usage, the tests are linked with `lld` now.

Around a minute of execution time is added for installing `grcov`,
another minute is added for extracting the data. The installation
could be modified to either happen in the image creation already, or to
download the precompiled release binary if required.

Closes: #468