jobs: implement data protection cleanup job for events
Introduce a new job with the identifier `event-cleanup`. It identifies
events that ended at least a defined number of days ago, and deletes
them including associated resources such as shared folders, and the room
if it is not connected to other events.

It accepts the following fields in its JSON parameters object:

- days_since_last_occurrence: Positive integer describing the number of
  days which must have passed since the meeting ended, default: 30.
- fail_on_shared_folder_deletion_error: Boolean flag which fails the
  cleanup job if the shared folder deletion on an external system such
  as NextCloud fails, default: false.

The existing `/internal/room/{room_id}` endpoint `DELETE` functionality
got refactored and reused for the room deletion part of the job. The new
event deletion implementation follows the same structure by implementing
the same trait called `Deleter` for `RoomDeleter` and `EventDeleter`.
This functionality lives in the newly added `opentalk-controller-utils`
crate.

Closes: #423