log: Add helper crate for writing to specific loggers
In order to write output from job execution to separate loggers which
can capture it, this crate introduces alternative variants of the
`log::{log,trace,debug,info,warn,error}` macros. These can accept an
additional `log: <logger>` parameter which will be used as the log
target.

This is considered an inadequate solution to fully capture every output,
because code paths could be executed inside a job which use the macros
from the `log` crate directly. It is good enough for the current
requirement of explicitly logging output from the job directly.