redis_args: add impl crate and use through wrapper crate
Closes #503

This decouples the usage of the `redis_args` crate from additional
imports that are required by it, such as `serde_json`, because they can
be depended upon and re-exported by the wrapper crate. This reduces
the amount of false positives for tools that try to detect superfluous
dependencies.

In order to keep consistency between the crate name that is referenced
by the macros and the name that is imported, I decided to rename the
`opentalk-redis-args` crate to `redis-args`. Referencing `redis_args`
inside the macros while naming the crate `opentalk_redis_args` in
reality is a mental burden that I'd like to avoid. I dislike using such
a generic name, but I think it's the better solution for this issue.