docs: add administration manual
See: #509

Closes: #510 - configuration file
Closes: #514 - command-line usage
Closes: #519 - HTTP service
Closes: #520 - jobs execution system
Closes: #529 - shared folder

This commit adds a new crate under `crates/ci-doc-updater` which can be
used to update source code blocks inside markdown files. For doing that,
it uses begin and end markers that are present in the file, and replaces
the text in between these markers. This is a very basic approach, and
its implementation is not too memory-efficient, but I expect it to be
robust enough for the use case, especially because the documentation
files are not very large. Through that mechanism, developers can get
notified by CI pipelines whenever any of the output changed and needs to
be updated inside the markdown documentation.

The `ci/docs-update-generated-parts.sh` script file was added. It
calls the `opentalk-controller` executable several times with different
command-line arguments in order to generate the output which is then
processed by the `ci-doc-updater` described above. Some of the `--help`
outputs generate trailing whitespace which is removed by a `sed` command
inside the script. For the script to work, certain services need to
be available, because for some subcommands such as `jobs execute`, the
controller connects to them. This is also reflected in the CI pipeline
changes.

I added a `--hide-duration` flag to the
`opentalk-controller jobs execute` subcommand which allows for
reproducible output, so that the CI does not repeatedly fail due to the
changes it detects in the output.

Some `TODO`s are present in markdown comments, almost always
accompanied by construction sign emojis. These will subsequently be
replaced in upcoming changes. It's better to have some of them in the
files instead of blocking the merge until the documentation is finished.