refactor: Introduce signaling module frontend data traits and collection structs
The new traits bind the namespace to the type that implements the
frontend data. They must be implemented by the frontend data and peer
frontend data of a signaling module which enforces their usage. If a
signaling module does not provide frontend data, the union type can be
used, it allows usage without a namespace.

The ModuleData and ModulePeerData structs contain sets of signaling
module frontend data for one participant.