docs(migration): fix release number and improve section for speaker detection configuration
The `v0.8.x` release series of the controller never saw a final release,
therefore the version number in the migration documentation should be
`0.9.0` because that's the first final release where this feature is
included.

Because the context might not be obvious to readers, I also added a link
to the `configuration` section in the migration documentation.