Fix: `DatabaseError` should map `diesel::result::Error::NotFound` to `DatabaseError::NotFound`
As you can see in the commit 57a55a1b9260c5adc9b7376c276bbb9afdbe8a77 ,
there was a `From` implementation for the `diesel::result::Error::NotFound`,
which was mapped to `DatabaseError::NotFound`. This was intended and was removed,
which shouldn't be the case.
This special case should be converted to a `DatabaseError::NotFound`

Closes #743