fix: specify usage of the serde feature for the opentalk-types dependency
Using the library with the `serde` feature in the workspace works just
fine when it is built in the workspace, even if the feature is not
consistently passed through to `opentalk-types`. It prevents publication
of the crate to crates.io though, and other users of the crate would
encounter the same problem.