fix(ci): ignore frontend api yaml file based on name
When ignoring with the full path, or even a path pattern like
`*/controller/frontend_api.yaml`, the file is checked none the less
if this repository is included in another repository as a submodule.
Switching to just the filename makes this work, and is probably specific
enough for our use case here.