refactor: move all crates into paths matching the crate name
This is a preparation for CI improvements which will enable usage of the
crate name variable to detect changes inside specific paths, allowing
to conditionally run tests only if a specific crate is affected. In
addition it removes some confusion about the meaning of folder names,
e.g. each signaling module can now be identified by the corresponding
path prefix.