refactor(settings): move rabbitmq and authz configuration to runtime settings
Because the Authz settings can only synchronize the controllers when a
RabbitMq pool is available, and this must be respected when creating the
Authz instance, these two sections must be migrated together.

A nice side effect of this change is that the RabbitMq configuration is
now truly optional. As a consequence, the mail service and the recorder
configuration are now fully dependent on the availability of RabbitMq,
so an attempt to send out mails will only be made if the service is
available, and also recording is only made available in the meeting when
RabbitMq is configured accordingly.

Closes #1018 #1020