fix(api): make paginated query work in async trait functions
The `Paginated::load_and_count(…)` method does not work in all
locations, it fails with the error message "Implementation of
`diesel::query_builder::Query` is not general enough error" when called
from inside an async trait method. Adjusting it like this makes it work
just fine in that location.

The requirement for this small change was discovered in a refactoring
towards abstracting the database access in a trait. It is compatible
with the current state of the code, and can be merged up-front.

Inspiration came from: https://stackoverflow.com/a/78948091